Time: 66 ms (51.28%), Space: 144.5 MB (12.82%) - LeetHub

+++ 3789-maximize-subarrays-after-removing-one-conflicting-pair/3789-maximize-subarrays-after-removing-one-conflicting-pair.java
@@ -9,7 +9,7 @@ class Solution {
         }        
         long ans = 0;
         int[] left = {0, 0};
-        int[] bonus = new int[n+1];
+        long[] bonus = new long[n + 1];
         for (int r=0;r<=n;r++) {
             for (int l:right[r]) {
                 if (l > left[0]) {
@@ -25,7 +25,7 @@ class Solution {
             }
         }        
         long maxBonus = 0;
-        for (int b : bonus) {
+        for (long b : bonus) {
             maxBonus = Math.max(maxBonus, b);
         }       
         return ans + maxBonus;
